+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Tactical Mind
 - Deflect Attacks
 - Heart of the Storm
+- Eldritch Invocations: Agonizing Blast
 ## API Update Notes
 - Added `sockets.syntheticActivityDataRoll` for remotely rolling modified activities, e.g. with changed damage.