+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.5.35 Release:
+## Update Notes:
+- This is not a V14 update. An update for V14 is still a work in progress.
+## Bug Fixes:
+- Iron Chin
+
 # 1.5.34 Release:
 ## Update Notes:
 - This is not a V14 update. An update for V14 is still a work in progress.